sem o inf

+++ PROJETO/ProjV1/projeto_taylor.c
@@ -6,7 +6,7 @@ int fatorial(int n){
     if (n == 0){
         return 1;
     }
-    int res = 1;
+    float res = 1;
     for (i = 1; i <= n; i++){
             res *= i;
         }
